transfer deprecated_method into framework module
