+++ volatility3/framework/configuration/requirements.py
@@ -11,7 +11,6 @@ expect to be in the context (such as particular layers or symboltables).
 import abc
 import logging
 import os
-import functools
 from typing import Any, ClassVar, Dict, List, Optional, Set, Tuple, Type
 from urllib import parse, request
 
@@ -724,25 +723,3 @@ class ModuleRequirement(
         """Builds the appropriate configuration for the specified
         requirement."""
         return context.modules[value].build_configuration()
-
-
-def deprecated_method(replacement: str, additional_information: str = ""):
-    """A decorator for marking functions as deprecated.
-
-    Args:
-        replacement: The replacement function overriding the deprecated API (full path preferred, starting from "volatility3."). String was preferred, for convenience and to prevent import conflicts on caller side.
-        additional_information: Information appended at the end of the deprecation message
-    """
-
-    def decorator(deprecated_func):
-        @functools.wraps(deprecated_func)
-        def wrapper(*args, **kwargs):
-            nonlocal replacement, additional_information
-            deprecation_msg = f"Method \"{deprecated_func.__module__ + '.' + deprecated_func.__name__}\" is deprecated, use \"{replacement}\" instead. {additional_information}"
-            vollog.warning(deprecation_msg)
-            # Return the wrapped function with its original arguments
-            return deprecated_func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
